100% clean build main.cpp and aiobject.cpp, target build now broken removed all unnessassary headers from headers, process of flattening namespace, mapping out dependencies moved global declairation from glland.h to main.h as comes from main.cpp promoted AIObj::statLethality from int to float. dev_notes for additional details

+++ aiobject.h
@@ -26,9 +26,9 @@
 #ifndef IS_AIOBJECT_H
 #define IS_AIOBJECT_H
 
-#include "common.h" // ok
-#include "model.h" // ok
-#include "effects.h" // ok
+// #include "common.h" // ok
+// #include "model.h" // ok
+// #include "effects.h" // ok
 
 // id values of objects
 // non-AI air objects
@@ -319,7 +319,7 @@ class AIObj : public DynamicObj
   int ttf; // time to fire missile, targeting mechanism
   CSmoke *smoke; // bright smoke behind the object (fighter&missiles)
   Uint32 timer;
-  int statLethality; // Lethality of offensive weapons
+  float statLethality; // Lethality of offensive weapons
   int DefensiveLethality; 
   bool dualshot; // one or two cannons?
   int manoeverstate; // changes to realistic manoevers.
